requirements.txt
absl-py==0.9.0
alembic==1.4.2
aniso8601==8.0.0
astroid==2.3.3
astunparse==1.6.3
attrs==19.3.0
botocore==1.17.3
CacheControl==0.12.6
cachetools==4.1.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
clickclick==1.2.2
colorama==0.4.3
connexion==2.7.0
coverage==5.1
docutils==0.16
ecdsa==0.15
entrypoints==0.3
flake8==3.7.9
Flask==1.1.2
Flask-Cors==3.0.8
Flask-JWT-Extended==3.24.1
Flask-Migrate==2.5.3
flask-restx==0.2.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask-Testing==0.8.0
gast==0.3.3
google-auth==1.17.2
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.28.1
gunicorn==20.0.4
h5py==2.10.0
httplib2==0.17.2
idna==2.9
importlib-metadata==1.6.0
inflection==0.5.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.9.5
jsonschema==3.2.0
Keras-Preprocessing==1.1.2
lazy-object-proxy==1.4.3
Mako==1.1.2
Markdown==3.2.2
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack==1.0.0
nose2==0.9.2
numpy==1.18.5
oauthlib==3.1.0
openapi-spec-validator==0.2.8
opt-einsum==3.2.1
passlib==1.7.2
protobuf==3.11.3
psycopg2==2.7.5
psycopg2-binary==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.20
pyflakes==2.2.0
PyJWT==1.7.1
pylint==2.4.4
pyrsistent==0.16.0
python-dateutil==2.8.1
python-dotenv==0.12.0
python-editor==1.0.4
python-jose==3.1.0
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
s3transfer==0.3.3
scipy==1.4.1
six==1.14.0
SQLAlchemy==1.3.16
tensorboard==2.2.2
tensorboard-plugin-wit==1.6.0.post3
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.8
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
